requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
arrow==1.3.0
asgiref==3.8.1
attrs==25.3.0
Authlib==1.6.0
binaryornot==0.4.4
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
cryptography==45.0.4
distro==1.9.0
Django==5.2.3
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastmcp==2.8.1
flet==0.28.3
flet-cli==0.28.3
flet-desktop==0.28.3
flet-web==0.28.3
google-auth==2.40.3
google-genai==1.20.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.9.4
mdurl==0.1.2
oauthlib==3.3.1
ollama==0.5.1
openai==1.87.0
openapi-pydantic==0.5.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyhub-llm==0.7.2
PyJWT==2.10.1
PyMuPDF==1.26.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-multipart==0.0.20
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==7.4.2
referencing==0.36.2
repath==0.9.0
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sse-starlette==2.3.6
starlette==0.47.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.1
typer==0.16.0
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
watchdog==4.0.2
watchfiles==1.1.0
websockets==15.0.1
